integration/requirements.txt
--index-url http://pypi.org

importlib-metadata==1.6.1
locustio==0.13.5
msgpack-python==0.5.6
pytest==3.8.2
allure-pytest==2.5.3
PyYAML==3.13
requests==2.20.1
requests-toolbelt==0.8.0
pyotp==2.2.7
beautifulsoup4==4.6.3
pyhamcrest==1.9.0
jsonschema==2.6.0
SQLAlchemy==1.2.14
boto3==1.9.44
pytest-rerunfailures==5.0
kafka-python==1.4.3
pytest-xdist==1.24.1
more-itertools==5.0.0
pluggy==0.13.1
setuptools==42.0.1
ulid==1.1
webdavclient3==0.14
waiting==1.4.1
zipp==1.0.0

# internal
paudit-qa==1.0.1